# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/streetview": {
      "get": {
        "summary": "Get Street View Image",
        "description": "Street View Image Metadata APIを使用して画像のメタデータを取得\n\nArgs:\n    latitude: 緯度\n    longitude: 経度\n    size: 画像サイズ\n\nReturns:\n    dict: メタデータ",
        "operationId": "get_street_view_image_streetview_get",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Latitude"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Longitude"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "600x300",
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Street View Image Streetview Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/route": {
      "get": {
        "summary": "Route",
        "description": "ルートを生成\n\nArgs:\n    current_lat: 現在の緯度\n    current_lng: 現在の経度\n    radius: 半径\n\nReturns:\n    RouteResponse: ルート情報",
        "operationId": "route_route_get",
        "parameters": [
          {
            "name": "currentLat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Currentlat"
            }
          },
          {
            "name": "currentLng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Currentlng"
            }
          },
          {
            "name": "radius",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Radius"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MidPoint": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "image_latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Latitude"
          },
          "image_longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Longitude"
          },
          "image_utf8": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Utf8"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "MidPoint",
        "description": "中間地点を表すモデル(画像情報を含む)"
      },
      "Point": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "Point",
        "description": "座標点を表すモデル"
      },
      "RouteResponse": {
        "properties": {
          "departure": {
            "$ref": "#/components/schemas/Point"
          },
          "destination": {
            "$ref": "#/components/schemas/MidPoint"
          },
          "midpoints": {
            "items": {
              "$ref": "#/components/schemas/MidPoint"
            },
            "type": "array",
            "title": "Midpoints"
          },
          "overview_polyline": {
            "type": "string",
            "title": "Overview Polyline"
          }
        },
        "type": "object",
        "required": [
          "departure",
          "destination",
          "midpoints",
          "overview_polyline"
        ],
        "title": "RouteResponse",
        "description": "ルート情報を表すモデル"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}